requirements.txt
setuptools~=65.5.1
requests~=2.31.0
beautifulsoup4~=4.12.2
pandas~=2.1.4
numpy~=1.26.2
plotly~=5.18.0
seaborn~=0.13.0
matplotlib~=3.8.2
scikit-learn~=1.3.2
psycopg2
pyyaml